# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RentKarRideKar API Gateway",
    "description": "Complete interactive API documentation for RentKarRideKar services including Auth, Vehicle, KYC, Order, Earnings, Referral, Wallet, Disputes, Ratings, Communication, and Admin services.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/cart/items": {
      "post": {
        "summary": "Add Vehicle to Cart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "vehicleId": {
                    "type": "string"
                  },
                  "pickupAddress": {
                    "type": "string"
                  },
                  "pickupLat": {
                    "type": "number"
                  },
                  "pickupLng": {
                    "type": "number"
                  },
                  "dropAddress": {
                    "type": "string"
                  },
                  "dropLat": {
                    "type": "number"
                  },
                  "dropLng": {
                    "type": "number"
                  },
                  "pickupDatetime": {
                    "type": "string"
                  },
                  "tripType": {
                    "type": "string"
                  },
                  "returnDatetime": {
                    "type": "string"
                  },
                  "totalDays": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "tags": [
          "Cart Service"
        ]
      }
    },
    "/cart": {
      "get": {
        "summary": "View Cart",
        "parameters": [
          {
            "name": "promoCode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Cart Service"
        ]
      },
      "delete": {
        "summary": "Clear Cart",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Cart Service"
        ]
      }
    },
    "/cart/items/{cartItemId}": {
      "delete": {
        "summary": "Remove Item from Cart",
        "parameters": [
          {
            "name": "cartItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Cart Service"
        ]
      }
    },
    "/orders": {
      "post": {
        "summary": "Place Order from Cart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "passengerNote": {
                    "type": "string"
                  },
                  "promoCode": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      },
      "get": {
        "summary": "List My Orders",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PASSENGER",
                "VEHICLE_OWNER",
                "DRIVER"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}": {
      "get": {
        "summary": "Get Order Details",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/vehicles/{vehicleId}/accept": {
      "post": {
        "summary": "Owner Accept Order",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "driverMode": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/vehicles/{vehicleId}/reject": {
      "post": {
        "summary": "Owner Reject Order",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/vehicles/{vehicleId}/remove-driver": {
      "post": {
        "summary": "Remove Assigned Driver",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/vehicles/{vehicleId}/driver-accept": {
      "post": {
        "summary": "Driver Accept Ride",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/vehicles/{vehicleId}/driver-reject": {
      "post": {
        "summary": "Driver Reject Ride",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/vehicles/{vehicleId}/arrive": {
      "post": {
        "summary": "Driver Arrived at Pickup",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/vehicles/{vehicleId}/verify-otp": {
      "post": {
        "summary": "Verify OTP to Start Trip",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "otp": {
                    "type": "string"
                  }
                },
                "required": [
                  "otp"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/vehicles/{vehicleId}/complete": {
      "post": {
        "summary": "Complete Vehicle Trip",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/complete": {
      "post": {
        "summary": "Internal System Complete Order",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/cancel": {
      "post": {
        "summary": "Passenger Cancel Order",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/vehicles/{vehicleId}/owner-cancel": {
      "post": {
        "summary": "Owner Cancel Vehicle Booking",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/confirm-payment": {
      "post": {
        "summary": "Confirm Advance Payment",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/location": {
      "post": {
        "summary": "Update Live Location Tracking",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  }
                },
                "required": [
                  "latitude",
                  "longitude"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/tracking": {
      "get": {
        "summary": "Get Trip Tracking Locations",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/vehicles/{vehicle_id}/assign-driver": {
      "post": {
        "summary": "Assign Trusted Driver to Booking (NEW 2.1)",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "driverId": {
                    "type": "string"
                  }
                },
                "required": [
                  "driverId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/vehicles/{vehicle_id}/otp": {
      "get": {
        "summary": "Get Booking Vehicle OTP (NEW 9.1)",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
    "/orders/{orderId}/vehicles/{vehicle_id}/pay-balance": {
      "post": {
        "summary": "Pay Trip Final Balance (NEW 9.2)",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "paymentMethod": {
                    "type": "string"
                  },
                  "useWalletCredit": {
                    "type": "boolean"
                  },
                  "walletCreditAmount": {
                    "type": "number"
                  }
                },
                "required": [
                  "paymentMethod"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
...
    "/orders/owner/earnings": {
      "get": {
        "summary": "Get Owner Earnings Summary (NEW 4.1)",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "DAILY",
                "WEEKLY",
                "MONTHLY",
                "LIFETIME"
              ]
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Order Service"
        ]
      }
    },
...
  }
}